chore(j-s): extract HighlightColorPicker into its own component
